Fix some issues.

+++ circuits/cpp/barretenberg/src/aztec/rollup/constants.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <stddef.h>
+#include <stdint.h>
 
 namespace rollup {
 
@@ -14,7 +15,7 @@ constexpr size_t TX_FEE_BIT_LENGTH = 254 - MAX_TXS_BIT_LENGTH;
 constexpr size_t NUM_ASSETS_BIT_LENGTH = 2;
 constexpr size_t NUM_ASSETS = 1 << NUM_ASSETS_BIT_LENGTH;
 
-constexpr size_t NUM_BRIDGE_CALLS_PER_BLOCK = 4;
+constexpr uint32_t NUM_BRIDGE_CALLS_PER_BLOCK = 4;
 
 namespace ProofIds {
 enum { JOIN_SPLIT = 0, ACCOUNT = 1, DEFI_DEPOSIT = 2, DEFI_CLAIM = 3 };
